+++ docs/news.md
@@ -5,8 +5,9 @@ This page documents notable changes introduced in each chronological release of 
 
 ## Version 3.2 (TBD)
 
-  * Removes JMetal library from distributions.  These added ~20 MBs to file sizes and are not required
-    for most uses.
+  * Removes JMetal as a required dependency at runtime.  Not only does this reduce the jar file size by
+    ~20 MBs, it removes overlapping class names shared by both libraries.
+
 
 ## Version 3.1 (19 Nov 2022)
